+++ src/render/figure.rs
@@ -527,6 +527,7 @@ fn clone_layout(l: &Layout) -> Layout {
     new.x_label_offset = l.x_label_offset;
     new.y_label_offset = l.y_label_offset;
     new.y2_label_offset = l.y2_label_offset;
+    new.scale = l.scale;
     new
 }